% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_Plots-Histogram.R
\name{histogram}
\alias{histogram}
\title{Histogram Plot Code}
\usage{
histogram(data = NULL, df_name = NULL, x = NULL, Theme = NULL,
  fontSize = 10, colourfill = "#00FF0080", colorby = "None",
  legendPos = "right", title_x = "", title_y = "", plotTitle = "",
  hideAxis, axisAngle, facetRow, facetCol)
}
\arguments{
\item{data}{data.frame - Passed from shiny enviornment}

\item{x}{character- This is x variable for aesthitic, supplied from shiny enviornment.}

\item{Theme}{character- parameter to change plot theme. (defaults to None).}

\item{colorby}{character - Parameter used to colorby plot. (defaults to None).}

\item{legendPos}{character- Adjust Legend position (defaults to right)}

\item{title_x}{character- Set x axis title.}

\item{title_y}{character- Set y axis title.}

\item{plotTitle}{character- Set plot title.}

\item{y}{character- This is y variable for aesthitic, supplied from shiny enviornment.}

\item{fontsize}{numeric- This parameter controls fontsize for x-y titles, x-y ticks labels. (defaults to 10).}
}
\value{
returns Plot and Code
}
\description{
Helper function to generate bar plot based on different conditions.
}
