% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Plots-Scatter.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Bar Plot Code}
\usage{
scatter_plot(data = dt, x = NULL, y = NULL, Theme = NULL,
  colourfill = "#00FF0080", colorby = "None", shapeby, fontSize = 10,
  legendPos = "right", dotSize = 2, dotOpa = 0.7, title_x = "",
  title_y = "", plotTitle = "", hideAxis, axisAngle,
  regressionLine = FALSE, correlation = FALSE, facetRow, facetCol)
}
\arguments{
\item{data}{data.frame - Passed from shiny enviornment}

\item{x}{character- This is x variable for aesthitic, supplied from shiny enviornment.}

\item{y}{character- This is y variable for aesthitic, supplied from shiny enviornment.}

\item{Theme}{character- parameter to change plot theme. (defaults to None).}

\item{colourfill}{character - Parameter used to colour plot.}

\item{colorby}{character - Parameter used to colorby plot. (defaults to None).}

\item{legendPos}{character- Adjust Legend position (defaults to right)}

\item{dotSize}{numeric- Set size of dots.}

\item{dotOpa}{numeric- Set opacity of dots.}

\item{title_x}{character- Set x axis title.}

\item{title_y}{character- Set y axis title.}

\item{plotTitle}{character- Set plot title.}

\item{regressionLine}{checkbox- to add regression line to plot.}

\item{correlation}{checkbox- to calculate pearson correlation.}

\item{fontsize}{numeric- This parameter controls fontsize for x-y titles, x-y ticks labels. (defaults to 10).}
}
\value{
returns H2OFrame
}
\description{
Helper function to generate bar plot based on different conditions.
}
