% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_Plots-Bar.R
\name{bar_plot}
\alias{bar_plot}
\title{Bar Plot Code}
\usage{
bar_plot(data = NULL, df_name = NULL, x = NULL, y = NULL,
  colorby = "None", Theme = NULL, fontSize = 10,
  legendPos = "right", title_x = "", title_y = "",
  colourfill = "#00FF0080", plotTitle = "", facetRow, facetCol,
  hideAxis, axisAngle = 90, position = "", coorflip = FALSE,
  interactive = FALSE)
}
\arguments{
\item{data}{data.frame - Passed from shiny enviornment}

\item{x}{character- This is x variable for aesthitic, supplied from shiny enviornment.}

\item{y}{character- This is y variable for aesthitic, supplied from shiny enviornment.}

\item{colorby}{character - Parameter used to colorby plot. (defaults to None).}

\item{Theme}{character- parameter to change plot theme. (defaults to None).}

\item{legendPos}{character- Adjust Legend position (defaults to right)}

\item{title_x}{character- Set x axis title.}

\item{title_y}{character- Set y axis title.}

\item{colourfill}{character- Hexcode or RGB color code for plot (defaults to #00FF0080)}

\item{plotTitle}{character- Set plot title.}

\item{facetRow}{character- facet column name, if faceted by Row}

\item{facetCol}{character- facet column name, if faceted by Column}

\item{hideAxis}{Logical - If we want to hide x Axis labels. (default: FALSE)}

\item{axisAngle}{Numeric- change x tick labels angle. (default: 90)}

\item{position}{character- Three Choice Stack, Fill, Dodge (Default: Stack)}

\item{coorflip}{Logical- Flip Coordinates (Defaults: FALSE)}

\item{interactive}{Logical- Convert to interactive plot (Defaults:FALSE)}

\item{fontsize}{numeric- This parameter controls fontsize for x-y titles, x-y ticks labels. (defaults to 10).}
}
\value{
returns H2OFrame
}
\description{
Helper function to generate bar plot based on different conditions.
}
